Clean instance records from registry (#127)
Delete instance from registry (if exists) during delete-instance, otherwise it still gets returned by `list-instances`, meaning it never appears to delete.

+++ source/Octopus.Shared/Configuration/Instances/RegistryApplicationInstanceStore.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Microsoft.Win32;
+using Octopus.Diagnostics;
 using Octopus.Shared.Util;
 
 namespace Octopus.Shared.Configuration.Instances
@@ -16,13 +17,15 @@ namespace Octopus.Shared.Configuration.Instances
         const string KeyName = "Software\\Octopus";
 
         readonly StartUpInstanceRequest startUpInstanceRequest;
+        readonly ILog log;
 
-        public RegistryApplicationInstanceStore(StartUpInstanceRequest startUpInstanceRequest)
+        public RegistryApplicationInstanceStore(StartUpInstanceRequest startUpInstanceRequest, ILog log)
         {
             this.startUpInstanceRequest = startUpInstanceRequest;
+            this.log = log;
         }
 
-        public ApplicationInstanceRecord GetInstanceFromRegistry(string instanceName)
+        public ApplicationInstanceRecord? GetInstanceFromRegistry(string instanceName)
         {
             var allInstances = GetListFromRegistry();
             return allInstances.SingleOrDefault(i => i.InstanceName.Equals(instanceName, StringComparison.CurrentCultureIgnoreCase));
@@ -32,53 +35,78 @@ namespace Octopus.Shared.Configuration.Instances
         {
             var results = new List<ApplicationInstanceRecord>();
 
-            if (PlatformDetection.IsRunningOnWindows)
-                using (var rootKey = RegistryKey.OpenBaseKey(Hive, View))
-                using (var subKey = rootKey.OpenSubKey(KeyName, false))
+            if (!PlatformDetection.IsRunningOnWindows)
+                return results;
+
+            using (var rootKey = RegistryKey.OpenBaseKey(Hive, View))
+            using (var subKey = rootKey.OpenSubKey(KeyName, false))
+            {
+                if (subKey == null)
+                    return results;
+
+                using (var applicationNameKey = subKey.OpenSubKey(startUpInstanceRequest.ApplicationName.ToString(), false))
                 {
-                    if (subKey == null)
+                    if (applicationNameKey == null)
                         return results;
 
-                    using (var applicationNameKey = subKey.OpenSubKey(startUpInstanceRequest.ApplicationName.ToString(), false))
+                    var instanceNames = applicationNameKey.GetSubKeyNames();
+
+                    foreach (var instanceName in instanceNames)
                     {
-                        if (applicationNameKey == null)
-                            return results;
+                        using (var instanceKey = applicationNameKey.OpenSubKey(instanceName, false))
+                        {
+                            if (instanceKey == null)
+                                continue;
 
-                        var instanceNames = applicationNameKey.GetSubKeyNames();
-
-                        foreach (var instanceName in instanceNames)
-                            using (var instanceKey = applicationNameKey.OpenSubKey(instanceName, false))
-                            {
-                                if (instanceKey == null)
-                                    continue;
-
-                                var path = instanceKey.GetValue("ConfigurationFilePath");
-                                results.Add(new ApplicationInstanceRecord(instanceName, (string)path));
-                            }
+                            var path = instanceKey.GetValue("ConfigurationFilePath");
+                            results.Add(new ApplicationInstanceRecord(instanceName, (string)path));
+                        }
                     }
                 }
+            }
 
             return results;
         }
 
         public void DeleteFromRegistry(string instanceName)
         {
-            if (PlatformDetection.IsRunningOnWindows)
-                using (var rootKey = RegistryKey.OpenBaseKey(Hive, View))
-                using (var subKey = rootKey.OpenSubKey(KeyName, true))
-                {
-                    if (subKey == null)
-                        return;
+            if (!PlatformDetection.IsRunningOnWindows)
+                return;
 
-                    using (var applicationNameKey = subKey.OpenSubKey(startUpInstanceRequest.ApplicationName.ToString(), true))
+            var registryInstance = GetInstanceFromRegistry(instanceName);
+            if (registryInstance == null)
+                return;
+
+            try
+            {
+                using (var rootKey = RegistryKey.OpenBaseKey(Hive, View))
+                {
+                    using (var subKey = rootKey.OpenSubKey(KeyName, true))
                     {
-                        if (applicationNameKey == null)
+                        if (subKey == null)
                             return;
 
-                        applicationNameKey.DeleteSubKey(instanceName);
-                        applicationNameKey.Flush();
+                        using (var applicationNameKey = subKey.OpenSubKey(startUpInstanceRequest.ApplicationName.ToString(), true))
+                        {
+                            if (applicationNameKey == null)
+                                return;
+
+                            applicationNameKey.DeleteSubKey(instanceName);
+                            applicationNameKey.Flush();
+                        }
                     }
                 }
+            }
+            catch (UnauthorizedAccessException ex)
+            {
+                log.Error(ex, $"Unable to delete instance '{instanceName}' from registry at {Hive}\\{KeyName}\\{startUpInstanceRequest.ApplicationName}\\{instanceName} as user '{Environment.UserName}'.");
+                throw new ControlledFailureException($"Unable to delete instance '{instanceName}' from registry as user '{Environment.UserName}'. Please check your permissions.");
+            }
+            catch (Exception ex)
+            {
+                log.Error(ex, $"Unable to delete instance '{instanceName}' from registry at {Hive}\\{KeyName}\\{startUpInstanceRequest.ApplicationName}\\{instanceName} as user '{Environment.UserName}'.");
+                throw new ControlledFailureException($"Unable to delete instance '{instanceName}' from registry as user '{Environment.UserName}'.");
+            }
         }
     }
 }
